api: remove sqlc (#41)
For the current stage of this project, sqlc is not the write tool. We
only have a few queries and so the abstractions isn't worth the cost.

This is ground work for a future PR that will make use of jsonb features
that sqlc hasn't quite mastered.

If in the future, we find the sqlc will add value, we can add it back
at that time.