api: store proofs on tree table (#40)
I don't think it is necessary to store the pre-computed proofs for a
tree in a separate table. In fact, the two tables design yields N
queries for a tree with N leaves. These queries are currently batched,
but I believe we can avoid the work altogether.

I did notice that the current code dealt with duplicate leaves by
selecting the first found proof. On further inspection, it's unclear
what behaviour we should exhibit on duplicate keys. We may want to
return an error on tree creation if we detect duplicate leaves.