fix: simplify data model, proofs_hashes table
historically, we've kept a column called `proofs` on the trees table. this is a jsonb column that contained all of the proofs for a tree.

at scale, this caused a couple of issues. the first thing that we saw is that it became too expensive to make a GIN index on proofs at the time of insert. so, we created the `trees_proofs` table which is a copy of the trees table, but replicated in the background with the index. this worked for a while. a couple of weeks ago, we noticed that the `trees` and `trees_proofs` table became out of date. the [background job](https://github.com/contextwtf/lanyard/blob/82f31584ac7f887057170b1b39315ef1380269be/cmd/api/main.go#L37-L54) to backfill the trees was just stalling & deadlocking. we tried a couple of fixes (#97 + #98), but neither worked. this had the effect of breaking the proofs -> roots endpoint for new trees.

simultaneously, we noticed that with larger trees of 200K+ addresses, the proofs column on `trees` was too large to even insert. the limit of a jsonb column is 255MB, which we were hitting.

this pull request fixes these issues by creating a new table that we write to at time of tree creation in a tx, called `proofs_hashes`. this table is very simple: [root, keccak256(proofs)]. we can efficiently query this when we want to look up roots for a given tree.

to make this work, we also have to update the code to generate a proof for a given hash. instead of querying the proof from the DB, we have to fetch the whole tree and recalculate it. the interesting thing is, for very large trees in production, this already takes 500ms when fetching from the db. testing in local dev yields response times of 200ms. this isn't perfect given that production doesn't run the db on the same instance, but i think it's safe to say that calculating the tree is not the bottleneck. will monitor after this is merged. also writes seem to be faster, because we're sending less data over the wire to postgres [no more 500MB json blobs! even for a 200K item hash, it's closer to 10-20MB]

- [ ] write migration script to backfill the `proofs_hashes` table
- after this is merged, a migration to delete the unneeded tables/data